integrated CAN ZPiezo
