+++ DeviceAdapters/WieneckeSinske/WieneckeSinske.vcxproj
@@ -167,11 +167,17 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="CAN29Axis.cpp" />
     <ClCompile Include="WieneckeSinske.cpp" />
+    <ClCompile Include="XYStageDevice.cpp" />
+    <ClCompile Include="ZPiezoCanDevice.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="CAN29.h" />
+    <ClInclude Include="CAN29Axis.h" />
     <ClInclude Include="WieneckeSinske.h" />
+    <ClInclude Include="XYStageDevice.h" />
+    <ClInclude Include="ZPiezoCanDevice.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\MMDevice\MMDevice-SharedRuntime.vcxproj">
