+++ DeviceAdapters/WieneckeSinske/WieneckeSinske.vcxproj.filters
@@ -21,12 +21,30 @@
     <ClCompile Include="CAN29.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="ZPiezoCanDevice.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="CAN29Axis.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="XYStageDevice.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="WieneckeSinske.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="CAN29.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="ZPiezoCanDevice.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="CAN29Axis.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="XYStageDevice.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
